+++ CHANGELOG.md
@@ -5,7 +5,8 @@ Full documentation for rocLibrary is available at [https://rocm.docs.amd.com/pro
 ## rocAL 2.3.0 (unreleased)
 
 ### Added
-* Extended support to rocAL's video decoder to use rocDecode hardware decoder.
+* Extended support to rocAL's video decoder to use rocDecode hardware decoder
+* Setup - installs rocdecode dev packages for Ubuntu, RedHat, and SLES
 
 ### Changed